implemented the new skel algorithm

+++ Filter/Binary.cs
@@ -6,9 +6,9 @@ namespace FingerprintRecognition.Filter {
 
     internal class Binary {
 
-        static public bool[,] Create(double[,] src, double threshold) {
+        static public bool[,] Create(byte[,] src, double threshold) {
             bool[,] res = new bool[src.GetLength(0), src.GetLength(1)];
-            MatTool<double>.Forward(ref src, (y, x, v) => {
+            MatTool<byte>.Forward(ref src, (y, x, v) => {
                 res[y, x] = v >= threshold;
                 return true;
             });
